+++ src/components/use/csrf.js
@@ -1,10 +1,10 @@
 import bootstrap from '../axios';
-import useBackend from './use-backend';
+import useRest from './rest';
 
 export default function useCsrf(tenant, wid) {
   bootstrap(tenant, wid);
 
-  const [csrf, csrfError, csrfIsLoading, csrfResponse] = useBackend({ url: '/v1/csrf-token' });
+  const [csrf, csrfError, csrfIsLoading, csrfResponse] = useRest({ url: '/v1/csrf-token' });
   const needsLogin = csrfError && csrfError.response && csrfError.response.status === 401;
   if (!csrfIsLoading && needsLogin) {
     window.location.assign(`${tenant}/login/?qlik-web-integration-id=${wid}&returnto=${window.location.href}`);
